+++ requirements.txt
@@ -1,6 +1,8 @@
 boto3==1.26.43
 botocore==1.29.43
 click==8.1.3
+DateTime==4.9
+duckdb==0.6.1
 dynamo-pandas==1.3.0
 Flask==2.2.2
 importlib-metadata==6.0.0
@@ -17,3 +19,4 @@ six==1.16.0
 urllib3==1.26.13
 Werkzeug==2.2.2
 zipp==3.11.0
+zope.interface==5.5.2
